requirements.txt
astroid==2.3.3
autopep8==1.4.4
backcall==0.1.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorama==0.4.1
cryptography==3.0
decorator==4.4.2
defusedxml==0.7.0rc1
diff-match-patch==20200713
Django==2.2.3
django-ckeditor==6.0.0
django-filter==2.4.0
django-import-export==2.4.0
django-js-asset==1.2.2
et-xmlfile==1.0.1
idna==2.10
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jdcal==1.4.1
jedi==0.16.0
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
mysql-connector-python==8.0.22
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
parso==0.6.2
pickleshare==0.7.5
Pillow==7.2.0
pipenv==2018.11.26
prompt-toolkit==3.0.4
protobuf==3.12.2
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rope==0.14.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
tablib==2.0.0
traitlets==4.3.3
urllib3==1.25.10
virtualenv==16.7.8
virtualenv-clone==0.5.3
wcwidth==0.1.8
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
